# cache the index of named parameters